Backend/requirements.txt
aniso8601==9.0.1
annotated-types==0.6.0
bcrypt==4.1.2
black==23.12.1
blinker==1.7.0
bunnet==1.2.0
click==8.1.7
defusedxml==0.7.1
dnspython==2.4.2
flake8==7.0.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Pydantic==0.12.0
Flask-RESTful==0.3.10
itsdangerous==2.1.2
Jinja2==3.1.3
lazy-model==0.2.0
MarkupSafe==2.1.3
mccabe==0.7.0
mongoengine==0.27.0
mypy-extensions==1.0.0
numpy==1.26.3
odfpy==1.4.1
packaging==23.2
pandas==2.2.0
pathspec==0.12.1
platformdirs==4.1.0
pycodestyle==2.11.1
pydantic==2.6.1
pydantic_core==2.16.2
pyflakes==3.2.0
PyJWT==2.8.0
pymongo==4.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
six==1.16.0
toml==0.10.2
typing_extensions==4.9.0
tzdata==2023.4
Werkzeug==3.0.1
xlrd==2.0.1
